+++ src/app/components/layers-permits/layers-permits.component.ts
@@ -4,6 +4,8 @@ import { UtilsService } from '../../services/utils.service';
 import { Router } from '@angular/router';
 import { environment } from 'src/environments/environment';
 import { Subject } from 'rxjs';
+import { MatDialog } from '@angular/material/dialog';
+import { DialogMessageComponent } from 'dist/sitmun-frontend-gui/';
 
 @Component({
   selector: 'app-layers-permits',
@@ -18,7 +20,8 @@ export class LayersPermitsComponent implements OnInit {
 
   permissionGroupTypes: Array<any> = [];
 
-  constructor(public cartographyGroupService: CartographyGroupService,
+  constructor(public dialog: MatDialog,
+    public cartographyGroupService: CartographyGroupService,
     private utils: UtilsService,
     private router: Router
   ) {
@@ -100,12 +103,22 @@ export class LayersPermitsComponent implements OnInit {
   }
 
   removeData(data: CartographyGroup[]) {
-    const promises: Promise<any>[] = [];
-    data.forEach(cartographyGroup => {
-      promises.push(new Promise((resolve, reject) => { this.cartographyGroupService.delete(cartographyGroup).toPromise().then((resp) => { resolve() }) }));
-      Promise.all(promises).then(() => {
-        this.dataUpdatedEvent.next(true);
-      });
+    const dialogRef = this.dialog.open(DialogMessageComponent);
+    dialogRef.componentInstance.title=this.utils.getTranslate("Caution");
+    dialogRef.componentInstance.message=this.utils.getTranslate("RemoveMessage");
+    dialogRef.afterClosed().subscribe(result => {
+      if(result){
+        if(result.event==='Accept') {  
+          const promises: Promise<any>[] = [];
+          data.forEach(cartographyGroup => {
+            promises.push(new Promise((resolve, reject) => { this.cartographyGroupService.delete(cartographyGroup).toPromise().then((resp) => { resolve() }) }));
+            Promise.all(promises).then(() => {
+              this.dataUpdatedEvent.next(true);
+            });
+          });
+      
+       }
+      }
     });
 
   }
